rename bbccouk.py -> bbc.py
